error when deleting comments

+++ app/controllers/api/comments_controller.rb
@@ -35,16 +35,17 @@ module Api
       c = Comment.where(id: params[:id]).first
       if (c)
         
-        Comment.where(id: params[:id]).destroy_all
+        
         reps = Report.where(id_comment: params[:id])
         reps.each do |item|
           Notification.where(idContent: item.id).where(isReports: true).destroy_all
         end
+        r = Report.where(id_comment: params[:id]).first
+        if (r)
+          Report.where(id_comment: params[:id]).destroy_all
+        end
 
-        Report.where(id_comment: params[:id]).destroy_all
-
-
-
+        Comment.where(id: params[:id]).destroy_all
         render json: { status: 'SUCCESS', message: 'ELIMINACION EXITOSA'}, status: :ok
       else
         render json: { status: 'INVALID', message: 'NO ENCONTRADA'}, status: :unauthorized
